+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## Fixed
+
+- renderWidget from WidgetConfigurationScreenProps does not update the widget
+
 ## [0.20.1] - 2026-01-15
 
 ## Fixed